bug fixed and add Api doc

+++ src/main/java/uz/mango/apps/fidotesttask/model/CreditDetails.java
@@ -1,25 +1,25 @@
 package uz.mango.apps.fidotesttask.model;
 
+import java.math.BigDecimal;
+
 public class CreditDetails {
 
     private int month;
-    private double totalSum;
-    private double interestPayment;
-    private double totalPayment;
-    private double remainingLoan;
+    private BigDecimal totalSum;
+    private BigDecimal interestPayment;
+    private BigDecimal totalPayment;
+    private BigDecimal remainingLoan;
 
-
-    public CreditDetails(int month, double principalPayment, double interestPayment, double totalPayment, double remainingLoan) {
+    public CreditDetails(int month, BigDecimal principalPayment, BigDecimal interestPayment, BigDecimal totalPayment, BigDecimal remainingLoan) {
         this.month = month;
         this.totalSum = principalPayment;
         this.interestPayment = interestPayment;
         this.totalPayment = totalPayment;
         this.remainingLoan = remainingLoan;
     }
-
     public int getMonth() { return month; }
-    public double getTotalSum() { return totalSum; }
-    public double getInterestPayment() { return interestPayment; }
-    public double getTotalPayment() { return totalPayment; }
-    public double getRemainingLoan() { return remainingLoan; }
+    public BigDecimal getTotalSum() { return totalSum; }
+    public BigDecimal getInterestPayment() { return interestPayment; }
+    public BigDecimal getTotalPayment() { return totalPayment; }
+    public BigDecimal getRemainingLoan() { return remainingLoan; }
 }
